Refactor button and card components to use new UI structure; remove old implementations
